Python analog of a tail command
Background:  Linux has a tail command that shows the last portion of a file.  Sometimes you want to re-invent the wheel.  Why not have a tail program of your own?  Seriously, Python has great native file manipulation capabilities built-in.  This program prints out the last characters of a file whereas Linux's tail is based on the last lines of a file.